+++ src/pages/admin/home.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { Button, Card, Col, Row } from 'react-bootstrap';
 import { useAppDispatch, useAppSelector } from '../../store/hooks';
-import { refreshInfo, selectInfo, selectInfoState } from '../../store/cr';
+import { refreshInfo, selectInfo, selectInfoState } from '../../store/store';
 import { Link } from 'react-router-dom';
 
 function HomePage(): JSX.Element {
