+++ src/pages/public/buy/index.tsx
@@ -3,7 +3,7 @@ import { Button, Form } from 'react-bootstrap';
 import { useHistory, useRouteMatch } from 'react-router-dom';
 import bt, { IBuyChannelRequest, IService } from '@synonymdev/blocktank-client';
 import { useAppDispatch, useAppSelector } from '../../../store/hooks';
-import { refreshInfo, selectInfo, selectInfoState } from '../../../store/cr';
+import { refreshInfo, selectInfo, selectInfoState } from '../../../store/store';
 import Spinner from '../../../components/spinner';
 import FormCard from '../../../components/form-card';
 import './index.scss';
